brain/requirements.txt
python-dotenv==1.1.1
mcp==1.10.1
openai==1.93.2
anthropic==0.57.1